requirements.txt
Flask
Flask-SocketIO
Flask-CORS
python-dotenv
azure-cognitiveservices-speech
gunicorn
gevent-websocket
